Feedback - MSI-X offsets back to 32-bits, add large bar test

+++ vmm_core/virt_whp/src/device.rs
@@ -379,10 +379,10 @@ impl AssignedPciDevice {
         }
     }
 
-    fn write_phys_config(&self, offset: u64, value: u32) {
+    fn write_phys_config(&self, offset: u16, value: u32) {
         match self.device.device().write_register(
             whp::abi::WHvVpciConfigSpace,
-            offset,
+            offset as u64,
             &value.to_ne_bytes(),
         ) {
             Ok(_) => (),
